[cross-gcc] fix dynamic linking problem with libstdc++
We must have the 'libstdc++.so' in both '$SYSROOT/usr/lib' and the
compiler's '$TARGET/lib' directory.

If missing in '$TARGET/lib' linking a C++ application with 'g++'
results in linking the 'libstdc++' statically.

If missing in '$SYSROOT/usr/lib' the 'ld' of newer binutils won't find
the libstdc++ under certain circumstances, e.g.: Linking a C application
with 'gcc' against a shared library implemented in C++.

With this patch we move the shared libs to '$SYSROOT/usr/lib' and create
links in '$TARGET/lib'. So that everybody shoud be happy.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>