[eglibc] Changed way to checkout a given subversion revision for eglibc
There are no release tarballs for eglibc. So I hacked the get stages
to use subversion to checkout a branch and revision. The code can
be moved to an external script and simplyfied/improved.

Functionality could be moved to usual make_get script in ptxdist, so
that other packages might profit from this feature.

Signed-off-by: Carsten Schlote <c.schlote@konzeptpark.de>