[eglibc-support] Integrated eglibc support into framework
Basically a clone of what is done for glibc. The source code must
be fetched from subversion for eglibc.

Right now I always update to HEAD. Which eventually results into updated
eglibc binaries each time you rebuild an eglibc toolchain.

Possible solution might be to add some extra field to nail the checkouts
to a given version. OTOH the release branches will be only for bugfixes, so
it worked always fine in the past.

Signed-off-by: Carsten Schlote <c.schlote@konzeptpark.de>