[build-script] Changed version detection from dirname to stored value
The VERSION variable is now extracted with git describe from
the latest annotated "OSELAS.Toolchain-<version>" tag. For untagged
working copies the '-x-y' postfix is stripped and a warning is
issued. This is a development feature. For releases tag the current
position with an annotated tag 'OSELAS.Toolchain-<version>'.

The required ptxdist version is taken from 'fixup-script.sh --info'.
The 'build_all_v2.mk update-ptxd' updates the submodule reference for
your current working copy.
The 'build_all_v2.mk compile-ptxd' can be used for testing and is
used by make to prepare the required ptxdist version for the
toolchain build.

Signed-off-by: Carsten Schlote <c.schlote@konzeptpark.de>